Move notifications to controller
